Update and rename functions.py to Python/armstrong.py
Moved to seperate file as suggested
